refactor: getErrorMessage.js파일을 util폴더로 변경
