+++ docs/README.md
@@ -33,10 +33,7 @@
   - [x] 게임이 끝난 후, 우승자를 알림창으로 확인할 수 있어야한다.
 
 ### 🔄 리팩터링 (todo)
-- [ ] isCarCanGo() 함수는 매번 결과값이 달라지기 때문에 예측이 불가능한 함수입니다. input 을 받아서 판단하는 로직으로 변경되면 좋을 거 같습니다. 그리고 isCarCanGo 요건 Car 모델이 있으니 거기서 메서드로 제공해주는 게 좋아 보입니당.
-- [ ] playRace 가 추가적으로 너무 많은 일들을 하고 있습니다.
-- [ ] 구조 (피드백)
-
-- [ ] 피드백 강의2 
+- [ ] 왈터님 피드백
 ## 질문 
 - [ ] import 그룹화... 너무 과한거 아닌가? 
+- [ ] isCarCanGo() 함수는 매번 결과값이 달라지기 때문에 예측이 불가능한 함수입니다. input 을 받아서 판단하는 로직으로 변경되면 좋을 거 같습니다. 그리고 isCarCanGo 요건 Car 모델이 있으니 거기서 메서드로 제공해주는 게 좋아 보입니당.(댓글 질문)
